fix(operation): stabilize Temporal adapter CLI
